+++ TaskweaveWidget/TaskweaveWidgetBundle.swift
@@ -1,9 +1,11 @@
-import WidgetKit
+import ActivityKit
 import SwiftUI
+import WidgetKit
 
 @main
 struct TaskweaveWidgetBundle: WidgetBundle {
     var body: some Widget {
         TaskweaveWidget()
+        TaskweaveLiveActivity()
     }
 }